Move structs to separate file.
