+++ metaqtv.go
@@ -23,54 +23,6 @@ import (
 	"time"
 )
 
-type Player struct {
-	Name    string
-	NameRaw []int
-	Team    string
-	TeamRaw []int
-	Colors  [2]int
-	Frags   int
-	Ping    int
-	Spec    bool
-	Time    int
-	IsBot   bool
-}
-
-type Spectator struct {
-	Name    string
-	NameRaw []int
-	IsBot   bool
-}
-
-type QTV struct {
-	Host          string
-	Address       string
-	Numspectators int
-	Spectators    []string
-}
-
-type Server struct {
-	IPAddress     string `json:"IpAddress"`
-	Address       string
-	Description   string
-	Title         string
-	Port          uint16
-	Country       string
-	Map           string
-	NumClients    int
-	MaxPlayers    int
-	NumPlayers    int
-	NumSpectators int
-	MaxSpectators int
-	Settings      map[string]string
-	QtvOnly       bool
-	Players       []Player
-	Spectators    []Spectator
-	QTV           []QTV
-
-	keepaliveCount int
-}
-
 type MutexStore struct {
 	sync.RWMutex
 	data []byte
@@ -333,6 +285,7 @@ func main() {
 						Port:           server.Port,
 						Settings:       map[string]string{},
 						Players:        make([]Player, 0),
+						Spectators:     make([]Spectator, 0),
 						QTV:            make([]QTV, 0),
 						keepaliveCount: keepalive,
 					}
